--- new_webapp/src/constants/strings.js
@@ -1,1 +1,0 @@
-export const baseRoute = process.env.REACT_APP_DEVELOPMENT ? "http://localhost:5000/" : "/"
